Migrate LinkedData to TS
